Add list of weeks, problem sets and links
Added information about the problem sets stored in the CS50-assignments
repository. Each problem set includes a corresponding link and is grouped
by week according the the CS50 curriculum. Additionally, provided the
homepage for the CS50 class with more information regarding the material
learned in the class, and the active learning assignments.